Replace all unscoped enums in interprt.h with constexpr and enum class
Modernisation contributing to #531

+++ Core/DolphinVM/decode.cpp
@@ -569,7 +569,7 @@ void HexDump(tracestream out, LPCTSTR lpszLine, uint8_t* pby,
 	ASSERT(nBytes > 0);
 	ASSERT(nWidth > 0);
 
-	int nRow = 0;
+	size_t nRow = 0;
 
 	tracestream::char_type oldFill = out.fill('0');
 	out << std::hex;
